Refactor: Loader 컴포넌트 폴더 위치 변경
/src/components/loader -> /src/components/state
