+++ FE/sidedish/src/components/state/Loader.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { LoadingDots } from "style/loader/loader";
+import { LoadingDots } from "style/state/loader";
 
 const Loader = () => <LoadingDots>Loading...</LoadingDots>;
